+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Direct installation and update support for raw HTTP(S) Python `.py` source URLs
+
 ## [1.6.0] - 2026-02-18
 
 ### Added